Inline invisible.vbs script
Currently, `windosu-elevate.cmd` and `windosu-runas.cmd` call a separate
file named `invisible.vbs` to perform their tasks.

This approach doesn't work when running Windosu in Electron as an `asar`
package. `asar` archives are mounted as virtual directories in Electron.

More information about `asar` archives here:

https://github.com/atom/electron/blob/master/docs/tutorial/application-packaging.md

Given that the files don't really exist in the filesystem, the call to
`invisible.vbs` fails. Electron copes with this limitation by patching
certain node fs functions like `child_process.execFile` to extract the
file before executing it, but this of course is not possible from a cmd
file.

A decent workaround, given that `invisible.vbs` is a one line script, is
to inline it in the scripts is being called.

Fixes: https://github.com/resin-io/herostratus/issues/20